+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### 🏷️ Changed/removed
 
 * Removed the `reset_lockout` argument for the shutdown recipe since it's not required to allow the dome to move.
+* Do not close the dome during daytime if the Overwatcher is disabled.
 
 
 ## 1.6.0 - January 26, 2025